pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackmerlin-solver"
version = "0.1.0"
description = "Automated hackmerlin.io puzzle solver using Browser Use"
authors = [{name = "Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "python-dotenv",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "venv*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
